refactor: move PerkCard to components/composite/cards
